requirements.txt
Django==3.2.13
biopython==1.79
rq==1.10.1
django-cors-headers==3.11.0
django-redis-cache==3.0.0
django-celery-beat==2.2.0
django-rq==2.5.1
psycopg2-binary==2.9.3
gunicorn==20.1.0 
requests==2.22.0
cryptography==36.0.2 